+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Changed
 
 - Unknown entry point error when calling a contract counts as a panic
+- Cairo edition set to `2023_11`
 
 #### Fixed